did some minor changes
