+++ example/index.html
@@ -48,7 +48,7 @@
           <label for="draw-tool">Tool: </label>
           <select name="draw-tool" id="draw-tool" value="brush">
             <option value="brush">Brush</option>
-            <option value="gradientBrush">Gradient Brush</option>
+            <option value="gradient_brush">Gradient Brush</option>
             <option value="eraser">Eraser</option>
             <option value="line">Line</option>
           </select>
